+++ Source/SimpleUECharts/Public/Data/ChartHover.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "CoreMinimal.h"
+
+DECLARE_DELEGATE_OneParam(FOnSlateChartHoverChanged, int32);