+++ Source/SimpleUECharts/Public/Slate/SPieChart.h
@@ -2,6 +2,7 @@
 
 #include "CoreMinimal.h"
 #include "Data/ChartDataPoint.h"
+#include "Data/ChartHover.h"
 #include "Data/ChartStyles.h"
 #include "Rendering/RenderingCommon.h"
 #include "Widgets/SLeafWidget.h"
@@ -10,6 +11,7 @@ class SIMPLEUECHARTS_API SPieChart : public SLeafWidget
 {
 public:
     SLATE_BEGIN_ARGS(SPieChart) {}
+        SLATE_EVENT(FOnSlateChartHoverChanged, OnHoveredDataPointChanged)
     SLATE_END_ARGS()
 
     void Construct(const FArguments& InArgs);
@@ -18,6 +20,7 @@ public:
     void ClearData();
     void SetChartStyle(const FChartStyle& InChartStyle);
     void SetPieChartStyle(const FPieChartStyle& InPieChartStyle);
+    int32 GetHoveredDataPointIndex() const;
 
     virtual FVector2D ComputeDesiredSize(float LayoutScaleMultiplier) const override;
 
@@ -29,6 +32,9 @@ public:
         int32 LayerId,
         const FWidgetStyle& InWidgetStyle,
         bool bParentEnabled) const override;
+
+    virtual FReply OnMouseMove(const FGeometry& MyGeometry, const FPointerEvent& MouseEvent) override;
+    virtual void OnMouseLeave(const FPointerEvent& MouseEvent) override;
 
 private:
     struct FPieSlice
@@ -57,12 +63,17 @@ private:
     void InvalidateCachedLayout();
     void EnsureCachedLayout(const FVector2D& LocalSize) const;
     FSlateFontInfo GetChartFont() const;
+    int32 FindHoveredDataPointIndex(const FVector2D& LocalPosition) const;
+    void SetHoveredDataPointIndex(int32 NewHoveredDataPointIndex);
+    void ClearHover();
 
     TArray<FChartDataPoint> Data;
     TArray<FPieSlice> CalculatedSlices;
     float TotalValue = 0.0f;
     FChartStyle ChartStyle;
     FPieChartStyle PieChartStyle;
+    int32 HoveredDataPointIndex = INDEX_NONE;
+    FOnSlateChartHoverChanged OnHoveredDataPointChanged;
     mutable bool bLayoutDirty = true;
     mutable FVector2D CachedLocalSize = FVector2D(-1.0f, -1.0f);
     mutable TArray<FSlateVertex> CachedVertices;