+++ Source/SimpleUECharts/Public/Widgets/BarChartWidget.h
@@ -43,6 +43,9 @@ public:
     UFUNCTION(BlueprintCallable, Category = "Chart")
     void RefreshChart();
 
+    UFUNCTION(BlueprintPure, Category = "Chart|Hover")
+    int32 GetHoveredDataPointIndex() const;
+
 protected:
     virtual TSharedRef<SWidget> RebuildWidget() override;
     virtual void SynchronizeProperties() override;
@@ -54,6 +57,8 @@ protected:
 
 private:
     void SynchronizeBarChartProperties();
+    void HandleSlateHoverChanged(int32 NewHoveredDataPointIndex);
 
     TSharedPtr<SBarChart> MyBarChart;
+    int32 HoveredDataPointIndex = INDEX_NONE;
 };